fix: create different file for success message component for better readability
