+++ packages/cfd/src/Containers/investor-password-manager/investor-password-manager.tsx
@@ -1,26 +1,11 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Field, Form, Formik, FieldProps } from 'formik';
-import { PasswordInput, PasswordMeter, Text, Button, Icon } from '@deriv/components';
+import { PasswordInput, PasswordMeter, Text, Button } from '@deriv/components';
 import { localize, Localize } from '@deriv/translations';
 import { getErrorMessages } from '@deriv/shared';
-import { TCFDPasswordSuccessMessage, TInvestorPasswordManager, TPasswordManagerModalFormValues } from '../props.types';
-
-const CFDPasswordSuccessMessage = ({ toggleModal, is_investor }: TCFDPasswordSuccessMessage) => (
-    <div className='cfd-password-manager__success'>
-        <Icon icon='IcPasswordUpdated' size={128} />
-        <Text as='p' size='xxs' align='center'>
-            {is_investor ? (
-                <Localize i18n_default_text='Your investor password has been changed.' />
-            ) : (
-                <Localize i18n_default_text='Your password has been changed.' />
-            )}
-        </Text>
-        <Button onClick={toggleModal} className='cfd-password-manager__success-btn' primary large>
-            <p className='dc-btn__text'>{localize('OK')}</p>
-        </Button>
-    </div>
-);
+import CFDPasswordSuccessMessage from './cfd-password-success-message';
+import { TInvestorPasswordManager, TPasswordManagerModalFormValues } from '../props.types';
 
 const InvestorPasswordManager = ({
     error_message_investor,
